examples/python/requirements.txt
azure-search-documents==11.6.0b9
